pyproject.toml
[project]
name = "travel-planner-decision-maker"
version = "0.1.0"
description = ""
authors = [
    {name = "Daniel James",email = "156598376+danieljamesMitb@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv (>=1.0.1,<2.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "googlemaps (>=4.10.0,<5.0.0)",
    "mysql-connector-python (>=9.2.0,<10.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "folium (>=0.19.5,<0.20.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
